test/topology: move topology rejoin test to own file
Move slow test for rejoining a server after a sudden stop to a separate
file.
