+++ test/topology/test_topology.py
@@ -277,12 +277,3 @@ async def test_remove_node_with_concurrent_ddl(manager, random_tables):
         logger.debug("ddl fiber done, finished")
 
 
-@pytest.mark.asyncio
-async def test_start_after_sudden_stop(manager: ManagerClient, random_tables) -> None:
-    """Tests a server can rejoin the cluster after being stopped suddenly"""
-    servers = await manager.running_servers()
-    table = await random_tables.add_table(ncolumns=5)
-    await manager.server_stop(servers[0].server_id)
-    await table.add_column()
-    await manager.server_start(servers[0].server_id)
-    await random_tables.verify_schema()
